requirements.txt
fastapi
uvicorn[standard]
requests
prometheus_client

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi